Backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "esannihith",email = "esanni04@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-chess (>=1.999,<2.0)",
    "python-socketio (>=5.13.0,<6.0.0)",
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn[standard] (>=0.35.0,<0.36.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
